+++ Player Accounts/Tests/Runtime/Utilities/UriHelperTests.cs
@@ -3,8 +3,14 @@ using UnityEngine;
 
 namespace Unity.Services.Authentication.PlayerAccounts.Tests
 {
+    /// <summary>
+    /// UriHelperTests
+    /// </summary>
     public static class UriHelperTests
     {
+        /// <summary>
+        /// Tests that UriHelper.ParseQueryString with a valid query works
+        /// </summary>
         [Test]
         public static void ParseQueryString_CalledWithValidQueryString_ReturnsCorrectDictionary()
         {
@@ -17,6 +23,9 @@ namespace Unity.Services.Authentication.PlayerAccounts.Tests
             Assert.AreEqual("value2", result["key2"]);
         }
 
+        /// <summary>
+        /// Tests that UriHelper.ParseQueryString with a null throws
+        /// </summary>
         [Test]
         public static void ParseQueryString_CalledWithNull_ThrowsException()
         {