+++ Runtime/IAuthenticationService.cs
@@ -47,20 +47,17 @@ namespace Unity.Services.Authentication
         /// Checks whether the player is signed in or not.
         /// A player can remain signed in but have an expired session.
         /// </summary>
-        /// <returns>Returns true if player is signed in, else false.</returns>
         bool IsSignedIn { get; }
 
         /// <summary>
         /// Checks whether the player is still authorized.
         /// A player is authorized as long as his access token remains valid.
         /// </summary>
-        /// <returns>Returns true if player is authorized, else false.</returns>
         bool IsAuthorized { get; }
 
         /// <summary>
         /// Checks whether the player session is expired.
         /// </summary>
-        /// <returns>Returns true if player's session expired.</returns>
         bool IsExpired { get; }
 
         /// <summary>
@@ -784,7 +781,6 @@ namespace Unity.Services.Authentication
         /// <param name="token">Unity Player Login's access token</param>
         /// <param name="options">Options for the operation</param>
         /// <returns>Task for the async operation</returns>
-        /// <returns>Task for the async operation</returns>
         /// <exception cref="AuthenticationException">
         /// The task fails with the exception when the task cannot complete successfully due to Authentication specific errors.
         /// <list type="bullet">
@@ -1064,7 +1060,7 @@ namespace Unity.Services.Authentication
         /// <param name="code">The sign-in code to be confirmed.</param>
         /// <param name="idProvider">The ID provider (optional).</param>
         /// <param name="externalToken">The external token (optional).</param>
-        /// <returns>A <see cref="Task"/> Representing the asynchronous operation. If the operation succeeds,
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation. If the operation succeeds,
         /// the task will complete successfully (indicating a successful response from the server).
         /// Otherwise, exceptions can be thrown as described below.</returns>
         /// <exception cref="AuthenticationException">Thrown when the code is null or empty.</exception>
@@ -1121,6 +1117,8 @@ namespace Unity.Services.Authentication
         /// <summary>
         /// Retrieves the Notifications that were created for the signed in player
         /// </summary>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.
+        /// The task result contains the list of <see cref="Notification"/>.</returns>
         /// <exception cref="AuthenticationException">
         /// The task fails with the exception when the task cannot complete successfully due to Authentication specific errors.
         /// <list type="bullet">