+++ Runtime/Models/Requests/Code Linking/CodeLinkInfoRequest.cs
@@ -1,11 +1,12 @@
+using System;
 using Newtonsoft.Json;
-using UnityEngine;
 
 namespace Unity.Services.Authentication
 {
+    [Serializable]
     class CodeLinkInfoRequest
     {
         [JsonProperty("signInCode")]
-        public string SignInCode { get; set; }
+        public string SignInCode;
     }
 }