+++ Runtime/Models/Requests/Code Linking/GenerateSignInCodeRequest.cs
@@ -1,5 +1,4 @@
 using System;
-using UnityEngine;
 using Newtonsoft.Json;
 
 namespace Unity.Services.Authentication